+++ R/markdown.R
@@ -93,3 +93,63 @@ escape_math = function(x) {
   x[i] = paste0(x[i], '`')
   x
 }
+
+#' Embed a file, multiple files, or directory on an HTML page
+#'
+#' For a file, first encode it into base64 data (a character string). Then
+#' generate a hyperlink of the form \samp{<a href="base64 data"
+#' download="filename">Download filename</a>}. The file can be downloaded when
+#' the link is clicked in modern web browsers. For a directory, it will be
+#' compressed as a zip archive first, and the zip file is passed to
+#' \code{embed_file()}. For multiple files, they are also compressed to a zip
+#' file first.
+#'
+#' These functions can be called in R code chunks in R Markdown documents with
+#' HTML output formats. You may embed an arbitrary file or directory in the HTML
+#' output file, so that readers of the HTML page can download it from the
+#' browser. A common use case is to embed data files for readers to download.
+#' @param path Path to the file(s) or directory.
+#' @param name The default filename to use when downloading the file. Note that
+#'   for \code{embed_dir()}, only the base name (of the zip filename) will be
+#'   used.
+#' @param text The text for the hyperlink.
+#' @param ... For \code{embed_file()}, additional arguments to be passed to
+#'   \code{htmltools::a()} (e.g., \code{class = 'foo'}). For \code{embed_dir()}
+#'   and \code{embed_files()}, arguments passed to \code{embed_file()}.
+#' @note Windows users may need to install Rtools to obtain the \command{zip}
+#'   command to use \code{embed_dir()} and \code{embed_files()}.
+#'
+#'   These functions require R packages \pkg{mime}, \pkg{base64enc}, and
+#'   \pkg{htmltools}. If you have installed the \pkg{rmarkdown} package, these
+#'   packages should be available, otherwise you need to install them
+#'   separately.
+#' @return An HTML tag \samp{<a>} with the appropriate attributes.
+#' @export
+#' @examples
+#' logo = file.path(R.home("doc"), "html", "logo.jpg")
+#' link = xfun::embed_file(logo, 'R-logo.jpg', 'Download R logo')
+#' link
+#' htmltools::browsable(link)
+embed_file = function(path, name = basename(path), text = paste('Download', name), ...) {
+  h = paste0("data:", mime::guess_type(path), ";base64,", base64enc::base64encode(path))
+  htmltools::a(text, href = h, download = name, ...)
+}
+
+#' @rdname embed_file
+#' @export
+embed_dir = function(path, name = paste0(normalize_path(path), '.zip'), ...) {
+  name  = gsub('/', '', basename(name))
+  in_dir(path, {
+    name = file.path(tempdir(), name); on.exit(file.remove(name), add = TRUE)
+    zip(name, '.'); embed_file(name, ...)
+  })
+}
+
+#' @rdname embed_file
+#' @export
+embed_files = function(path, name = with_ext(basename(path[1]), '.zip'), ...) {
+  name = file.path(tempdir(), basename(name))
+  on.exit(file.remove(name), add = TRUE)
+  zip(name, path)
+  embed_file(name, ...)
+}