+++ R/utils.R
@@ -11,6 +11,31 @@
 #' xfun::attr(z, 'foo')  # 2
 attr = function(...) base::attr(..., exact = TRUE)
 
+#' Set the global option \code{\link{options}(stringsAsFactors = FALSE)} inside
+#' a parent function and restore the option after the parent function exits
+#'
+#' This is a shorthand of \code{opts = options(stringsAsFactors = FALSE);
+#' on.exit(options(opts), add = TRUE)}; \code{strings_please()} is an alias of
+#' \code{stringsAsStrings()}.
+#' @export
+#' @examples
+#' f = function() {
+#' xfun::strings_please()
+#' data.frame(x = letters[1:4], y = factor(letters[1:4]))
+#' }
+#' str(f())  # the first column should be character
+stringsAsStrings = function() {
+  opts = options(stringsAsFactors = FALSE)
+  do.call(
+    on.exit, list(substitute(options(x), list(x = opts)), add = TRUE),
+    envir = parent.frame()
+  )
+}
+
+#' @rdname stringsAsStrings
+#' @export
+strings_please = stringsAsStrings
+
 #' Evaluate an expression under a specified working directory
 #'
 #' Change the working directory, evaluate the expression, and restore the