+++ tests/testit/test-utils.R
@@ -11,3 +11,11 @@ assert('in_dir() preserves the working directory', {
   in_dir('.', setwd(tempdir()))
   (same_path(owd, getwd()))
 })
+
+assert('stringsAsFactors() makes sure strings are not converted to factors', {
+  f = function() {
+    strings_please()
+    data.frame(x = letters[1:4], y = factor(letters[1:4]))
+  }
+  is.character(f()[, 1])
+})