+++ NEWS.md
@@ -1,3 +1,7 @@
+# CHANGES IN xfun VERSION 0.5
+
+
+
 # CHANGES IN xfun VERSION 0.4
 
 ## NEW FEATURES